+++ src/Model/Money.java
@@ -1,18 +1,15 @@
 package Model;
 
-
-import Model.Currency;
-
 public class Money {
-private final int amount;
+private final double amount;
 private final Currency currency;
 
-    public Money(int amount, Currency currency) {
+    public Money(double amount, Currency currency) {
         this.amount = amount;
         this.currency = currency;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
